gaming-signal-bot/pyproject.toml
[project]
name = "gaming-signal-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.15",
    "discord-py>=2.5.2",
    "python-dotenv>=1.1.1",
]
